+++ internal/validate/install/config.go
@@ -52,6 +52,16 @@ type Insight struct {
 	DeleteWhenDone bool             `yaml:"deleteWhenDone"`
 }
 
+type Executor struct {
+	Enabled bool `yaml:"enabled"`
+	Count   bool `yaml:"count"`
+}
+
+type Smtp struct {
+	Enabled bool   `yaml:"enabled"`
+	To      string `yaml:"to"`
+}
+
 type ValidationSpec struct {
 	// Search queries used for validation testing, e.g. "repo:^github\\.com/gorilla/mux$ Router".
 	SearchQuery []string `yaml:"searchQuery"`
@@ -61,6 +71,12 @@ type ValidationSpec struct {
 
 	// Insight used for validation testing.
 	Insight Insight `yaml:"insight"`
+
+	// Executor check configuration
+	Executor Executor `yaml:"executor"`
+
+	//Test SMTP configuration
+	Smtp Smtp `yaml:"smtp"`
 }
 
 // DefaultConfig returns a default configuration to be used for testing.
@@ -107,6 +123,13 @@ func DefaultConfig() *ValidationSpec {
 				},
 			},
 			DeleteWhenDone: true,
+		},
+		Executor: Executor{
+			Enabled: false,
+		},
+		Smtp: Smtp{
+			Enabled: false,
+			To:      "example@domain.com",
 		},
 	}
 }